Update migration operations

+++ src/models/users.model.js
@@ -9,7 +9,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: DataTypes.INTEGER
+        type: DataTypes.INTEGER(11)
       },
       password: {
         allowNull: false,
